IsMobile logic from InfoAboutBirth transferred to custom hook
