+++ src/common/tiles/PersonDetailsTile/InfoAboutBirth/index.js
@@ -1,19 +1,8 @@
-import { useEffect, useState } from "react";
-import { theme } from "../../../../theme";
 import { Wrapper, Content, NameItemDisplay, ContentItem } from "../../MovieDetailsTile/ProductionInfo/styled";
+import { useIsMobile } from "../../../../useIsMobile";
 
 export const InfoAboutBirth = ({ info, personDetailsSize }) => {
-    const [isMobile, setIsMobile] = useState(window.innerWidth <= theme.breakpoint.mobileMax);
-
-    useEffect(() => {
-        const handleResize = () => {
-            setIsMobile(window.innerWidth <= theme.breakpoint.mobileMax);
-        };
-
-        window.addEventListener('resize', handleResize);
-        return () => window.removeEventListener('resize', handleResize);
-    }, []);
-
+   const isMobile = useIsMobile();
 
     return (
         <Wrapper>
